feat(): move bind/ast to contract package
