+++ contract/ast/contract.go
