+++ contract/ast/entrypoint.go
