+++ contract/bind/interface.go
