+++ contract/bind/map.go
