+++ contract/bind/or.go
