+++ contract/bind/or_test.go
