+++ internal/generate/contract.go.tmpl
@@ -22,12 +22,11 @@ import (
 	"time"
 
 	"blockwatch.cc/tzgo/contract"
+	"blockwatch.cc/tzgo/contract/bind"
 	"blockwatch.cc/tzgo/micheline"
 	"blockwatch.cc/tzgo/rpc"
 	"blockwatch.cc/tzgo/tezos"
 	"github.com/pkg/errors"
-
-	"blockwatch.cc/tzgo/bind"
 )
 
 {{$contract := pascal .Contract.Name}}
