+++ internal/parse/entrypoints.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 	"strconv"
 
-	"blockwatch.cc/tzgo/ast"
+	"blockwatch.cc/tzgo/contract/ast"
 	"blockwatch.cc/tzgo/micheline"
 	"github.com/pkg/errors"
 )
